Add files via upload
The objective of this practical is to develop a Python program that 
efficiently converts temperatures between Fahrenheit and Celsius scales. The program 
should allow users to input a temperature in either Fahrenheit or Celsius and then 
output the corresponding temperature in the other scale. This exercise will help 
students understand basic programming concepts like input/output operations, data 
types, and arithmetic operations in Python.